[REFACTOR] main페이지 index.js로 옮김
